Move loggerMock and setWindowValue / resetWindowValue to single testHelper file
