+++ src/app/lib/utilities/preprocessor/index.test.js
@@ -1,4 +1,4 @@
-import loggerMock from '../../../../testHelpers/loggerMock';
+import { loggerMock } from '../../../../testHelpers/index';
 import preprocess from '.';
 
 const fixtureData = { foo: 'bar' };
