+++ src/testHelpers/index.js
@@ -1,6 +1,8 @@
 import renderer from 'react-test-renderer';
 import 'jest-styled-components';
 import ShallowRenderer from 'react-test-renderer/shallow';
+import nodeLogger from '../app/lib/logger.node';
+import deepClone from '../app/lib/json/deepClone';
 
 export const shouldMatchSnapshot = (title, component) => {
   it(title, () => {
@@ -30,3 +32,43 @@ export const isNull = (title, component) => {
     expect(tree).toBeNull();
   });
 };
+
+export const setWindowValue = (key, value) => {
+  const windowValue = window[key];
+  delete window[key];
+
+  let newValue = value;
+
+  if (value && typeof value === 'object') {
+    newValue = {
+      ...deepClone(windowValue),
+      ...value,
+    };
+  }
+
+  Object.defineProperty(window, key, {
+    value: newValue,
+    writable: true,
+  });
+};
+
+export const resetWindowValue = (key, value) => {
+  Object.defineProperty(window, key, {
+    value,
+    writable: true,
+  });
+};
+
+const mocks = {
+  error: jest.fn(),
+  warn: jest.fn(),
+  info: jest.fn(),
+  verbose: jest.fn(),
+  debug: jest.fn(),
+  silly: jest.fn(),
+};
+
+jest.mock('../app/lib/logger.node', () => jest.fn());
+nodeLogger.mockImplementation(() => mocks);
+
+export const loggerMock = mocks;
