--- src/testHelpers/setWindowValue.js
@@ -1,27 +1,0 @@
-import deepClone from '../app/lib/json/deepClone';
-
-export const setWindowValue = (key, value) => {
-  const windowValue = window[key];
-  delete window[key];
-
-  let newValue = value;
-
-  if (value && typeof value === 'object') {
-    newValue = {
-      ...deepClone(windowValue),
-      ...value,
-    };
-  }
-
-  Object.defineProperty(window, key, {
-    value: newValue,
-    writable: true,
-  });
-};
-
-export const resetWindowValue = (key, value) => {
-  Object.defineProperty(window, key, {
-    value,
-    writable: true,
-  });
-};
